chore:update 2020-01

+++ src/apps/src/main/java/com/x3platform/apps/services/impl/ApplicationMenuServiceImpl.java
@@ -1,22 +1,22 @@
 package com.x3platform.apps.services.impl;
 
-import com.alibaba.fastjson.JSON;
+import static com.x3platform.apps.Constants.APPLICATION_MENU_ROOT_ID;
+import static com.x3platform.apps.configuration.AppsConfiguration.APPLICATION_NAME;
+
+import com.x3platform.AuthorizationScope;
 import com.x3platform.KernelContext;
 import com.x3platform.apps.AppsSecurity;
-import com.x3platform.apps.configuration.AppsConfiguration;
 import com.x3platform.apps.mappers.ApplicationMenuMapper;
 import com.x3platform.apps.models.ApplicationMenu;
 import com.x3platform.apps.models.ApplicationMenuLite;
-import com.x3platform.apps.models.ApplicationMenuScopeInfo;
-import com.x3platform.apps.models.ApplicationMenuViewInfo;
 import com.x3platform.apps.services.ApplicationMenuService;
+import com.x3platform.cachebuffer.CachingManager;
 import com.x3platform.data.DataQuery;
 import com.x3platform.membership.Account;
+import com.x3platform.membership.MembershipManagement;
 import com.x3platform.util.StringUtil;
 import com.x3platform.util.UUIDUtil;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import org.springframework.beans.factory.annotation.Autowired;
 
 /**
@@ -24,11 +24,41 @@ import org.springframework.beans.factory.annotation.Autowired;
  */
 public class ApplicationMenuServiceImpl implements ApplicationMenuService {
 
+  private static final String CACHE_KEY_ID_PREFIX = "x3platform:apps:application-menu:id:";
+
+  private static final String DATA_SCOPE_TABLE_NAME = "application_menu_scope";
+
   /**
    * 数据提供器
    */
   @Autowired
   private ApplicationMenuMapper provider = null;
+
+  // -------------------------------------------------------
+  // 缓存管理
+  // -------------------------------------------------------
+
+  /**
+   * 添加缓存项
+   */
+  private void addCacheItem(ApplicationMenu item) {
+    if (!StringUtil.isNullOrEmpty(item.getId())) {
+      String key = CACHE_KEY_ID_PREFIX + item.getId();
+      CachingManager.set(key, item);
+    }
+  }
+
+  /**
+   * 移除缓存项
+   */
+  private void removeCacheItem(ApplicationMenu item) {
+    if (!StringUtil.isNullOrEmpty(item.getId())) {
+      String key = CACHE_KEY_ID_PREFIX + item.getId();
+      if (CachingManager.contains(key)) {
+        CachingManager.delete(key);
+      }
+    }
+  }
 
   // -------------------------------------------------------
   // 保存 删除
@@ -51,33 +81,23 @@ public class ApplicationMenuServiceImpl implements ApplicationMenuService {
 
     // 计算完整路径
     entity.setFullPath(combineFullPath(entity));
+
     if (provider.selectByPrimaryKey(id) == null) {
       affectedRows = provider.insert(entity);
     } else {
       affectedRows = provider.updateByPrimaryKey(entity);
     }
 
-    // [{},{},{},{}]
-    String authorazitionScope = entity.getAuthorizationReadScopeObjectText();
-    // 保存完整 菜单权限范围
-    if (!StringUtil.isNullOrEmpty(authorazitionScope)) {
-      List<ApplicationMenuScopeInfo> scopeInfos = JSON
-        .parseArray(entity.getAuthorizationReadScopeObjectText(), ApplicationMenuScopeInfo.class);
-      if (scopeInfos != null && scopeInfos.size() > 0) {
-        for (ApplicationMenuScopeInfo scopeInfo : scopeInfos) {
-          scopeInfo.setEntityId(id);
-          // authority_id
-          String authorityId = scopeInfo.getAuthorityId();
-          if (authorityId == null) {
-            // 设置默认可访问权限
-            scopeInfo.setAuthorityId("1-0003");
-          }
-          provider.addAuthorizationScopeObject(scopeInfo);
-        }
-      }
-    }
+    KernelContext.getLog().debug("save entity id:'{}', affectedRows:{}", id, affectedRows);
 
-    KernelContext.getLog().debug("save entity id:'" + id + "', affectedRows:" + affectedRows);
+    // 保存数据后, 更新缓存信息
+    entity = provider.selectByPrimaryKey(entity.getId());
+
+    if (entity != null) {
+      removeCacheItem(entity);
+
+      addCacheItem(entity);
+    }
 
     return 0;
   }
@@ -89,9 +109,17 @@ public class ApplicationMenuServiceImpl implements ApplicationMenuService {
    */
   @Override
   public int delete(String id) {
-    int affectedRows = provider.deleteByPrimaryKey(id);
+    ApplicationMenu entity = provider.selectByPrimaryKey(id);
 
-    KernelContext.getLog().debug("delete entity id:'" + id + "', affectedRows:" + affectedRows);
+    if (entity != null) {
+      // 删除缓存记录
+      removeCacheItem(entity);
+
+      // 删除数据库记录
+      int affectedRows = provider.deleteByPrimaryKey(id);
+
+      KernelContext.getLog().debug("delete entity id:'{}', affectedRows:{}", id, affectedRows);
+    }
 
     return 0;
   }
@@ -107,7 +135,29 @@ public class ApplicationMenuServiceImpl implements ApplicationMenuService {
    */
   @Override
   public ApplicationMenu findOne(String id) {
-    return provider.selectByPrimaryKey(id);
+    ApplicationMenu entity = null;
+
+    // 根节点快速处理
+    if (APPLICATION_MENU_ROOT_ID.equals(id)) {
+      return null;
+    }
+
+    String key = CACHE_KEY_ID_PREFIX + id;
+
+    if (CachingManager.contains(key)) {
+      entity = (ApplicationMenu) CachingManager.get(key);
+    }
+
+    // 如果缓存中未找到相关数据，则查找数据库内容
+    if (entity == null) {
+      entity = provider.selectByPrimaryKey(id);
+
+      if (entity != null) {
+        addCacheItem(entity);
+      }
+    }
+
+    return entity;
   }
 
   /**
@@ -120,25 +170,20 @@ public class ApplicationMenuServiceImpl implements ApplicationMenuService {
     return provider.findAll(query.getMap());
   }
 
-  @Override
-  public List<ApplicationMenuViewInfo> findApplicationMenuViewInfoAll(DataQuery query) {
-    return provider.findApplicationMenuViewInfoAll(query.getMap());
-  }
-
   /**
    * 查询所有相关记录
    *
    * @param query 数据查询参数
    */
   @Override
-  public List<ApplicationMenuLite> findAllQueryObject(DataQuery query) {
+  public List<ApplicationMenuLite> findAllLites(DataQuery query) {
 //     // 验证管理员身份
 //    if (AppsSecurity.IsAdministrator(KernelContext.Current.User, AppsConfiguration.ApplicationName)) {
 //      return this.provider.findAllQueryObject(whereClause, length);
 //    } else {
 //      return this.provider.findAllQueryObject(this.BindAuthorizationScopeSQL(whereClause), length);
 //    }
-    return provider.findAllQueryObject(query.getMap());
+    return provider.findAllLites(query.getMap());
   }
 
   // -------------------------------------------------------
@@ -197,13 +242,11 @@ public class ApplicationMenuServiceImpl implements ApplicationMenuService {
    * @param applicationId 所属应用标识
    * @param parentId 所属父级菜单标识
    * @param menuType 菜单类型
-   * @return 返回所有实例ApplicationMenu的详细信息
+   * @return 所有实例 {@link ApplicationMenu} 的详细信息
    */
   @Override
-  public List<ApplicationMenu> getMenusByParentId(
-    String applicationId, String parentId, String menuType) {
-    return provider.getMenusByParentId(applicationId, parentId, menuType,
-      bindAuthorizationScopeSql());
+  public List<ApplicationMenu> getMenusByParentId(String applicationId, String parentId, String menuType) {
+    return provider.getMenusByParentId(applicationId, parentId, menuType, bindAuthorizationScopeSql());
   }
 
   /**
@@ -245,7 +288,12 @@ public class ApplicationMenuServiceImpl implements ApplicationMenuService {
    */
   @Override
   public boolean hasAuthority(String entityId, String authorityName, Account account) {
-    return provider.hasAuthority(entityId, authorityName, account);
+    return MembershipManagement.getInstance().getAuthorizationObjectService().hasAuthority(
+      DATA_SCOPE_TABLE_NAME,
+      entityId,
+      KernelContext.parseObjectType(ApplicationMenu.class),
+      authorityName,
+      account);
   }
 
   /**
@@ -256,115 +304,49 @@ public class ApplicationMenuServiceImpl implements ApplicationMenuService {
    * @param scopeText 权限范围的文本
    */
   @Override
-  public void bindAuthorizationScopeObjects(String entityId, String authorityName,
-    String scopeText) {
-    provider.bindAuthorizationScopeObjects(entityId, authorityName, scopeText);
+  public void bindAuthorizationScopeByEntityId(String entityId, String authorityName, String scopeText) {
+    MembershipManagement.getInstance().getAuthorizationObjectService().bindAuthorizationScopeByEntityId(
+      DATA_SCOPE_TABLE_NAME,
+      entityId,
+      KernelContext.parseObjectType(ApplicationMenu.class),
+      authorityName,
+      scopeText);
   }
 
   /**
-   * 根据角色绑定菜单
+   * 配置应用菜单的权限信息
    *
-   * @param roleId 角色id
-   * @param bindScopeMenuList 字符串组合
-   * @param unBindScopeMenuList 解绑字符串组合
-   */
-  @Override
-  public void bindAuthorizationScopeByRole(String roleId, List<String> bindScopeMenuList,
-    List<String> unBindScopeMenuList) {
-    if (unBindScopeMenuList != null && unBindScopeMenuList.size() > 0) {
-      for (String scope : unBindScopeMenuList) {
-        String[] scopeKeys = scope.split("#");
-        provider.deleteAuthorizationScopeObjects(scopeKeys[4],
-          "com.x3platform.membership.models.RoleInfo", "2-0001", "Role", roleId);
-      }
-    }
-    // 判断当前是否存在 ， 如果不存在则存入， 删除的则删除
-    if (bindScopeMenuList != null && bindScopeMenuList.size() > 0) {
-      for (String scope : bindScopeMenuList) {
-        String[] scopeKeys = scope.split("#");
-        // 判断当前是否存在 如果存在,则不进行操作
-        boolean isExits = provider
-          .isExistAuthorizationScope(scopeKeys[4], "com.x3platform.membership.models.RoleInfo",
-            "2-0001", "Role", roleId);
-        if (!isExits) {
-          provider.bindAuthorizationScope(scopeKeys[4], "com.x3platform.membership.models.RoleInfo",
-            "2-0001", "Role", roleId);
-        }
-      }
-    }
-  }
-
-  /**
-   * 移除当前角色所有的菜单 ， 请注意在多个项目的情况下，统一控制和删除 ；
-   *
-   * @param roleId 角色id
-   * @param bindScopeList 新增绑定
-   */
-  @Override
-  public void bindMenuScopeByRole(String roleId, List<String> bindScopeList) {
-    // 删除全部 再插入
-    if (bindScopeList != null && bindScopeList.size() > 0) {
-      // 移除当前角色所有的菜单
-      provider.removeMenuScopeByRole(roleId); // 删除全部
-      for (int i = 0; i < bindScopeList.size(); i++) {
-        provider
-          .bindAuthorizationScope(bindScopeList.get(i), "com.x3platform.membership.models.RoleInfo",
-            "2-0001", "Role", roleId);
-      }
-    }
-  }
-
-  /**
-   * @param value 根据 value 获取 当前应用范围
-   */
-  @Override
-  public List<ApplicationMenuScopeInfo> getApplicationMenuScope(String value) {
-    List<ApplicationMenuScopeInfo> applicationMenuScopeInfos = new ArrayList<>();
-    if (StringUtil.isNullOrEmpty(value)) {
-      applicationMenuScopeInfos = JSON.parseArray(value, ApplicationMenuScopeInfo.class);
-    }
-    return applicationMenuScopeInfos;
-  }
-
-  /**
-   * @param applicationId 所属应用
-   * @param parentId 父级菜单
-   * @param menuType 所属菜单
-   */
-  @Override
-  public List<ApplicationMenu> getMenusScopeByParentId(String applicationId, String parentId,
-    String menuType, String accountId) {
-    return provider.getMenusScopeByParentId(applicationId, parentId, menuType, accountId);
-  }
-
-  /**
-   * 根据菜单 和 角色 查询是否存在
-   *
-   * @param roleId 角色id
-   * @param menuId 菜单id
-   */
-  @Override
-  public boolean isExistMenusScopeByRoleIdAndMenuId(String roleId, String menuId) {
-    String entityId = menuId;
-    String authorizationObjectId = roleId;
-    String authorizationObjectType = "Role";
-    return provider.isExistScope(authorizationObjectId, authorizationObjectType, entityId);
-  }
-
-  ///#endregion
-
-  ///#region 函数:GetAuthorizationScopeObjects(string entityId, string authorityName)
-
-  /**
-   * 查询应用菜单的权限信息
-   * @param entityId      实体标识
+   * @param authorizationObjectType 授权对象类型
+   * @param authorizationObjectId 授权对象标识
    * @param authorityName 权限名称
-   * @return
+   * @param entityIds 实体标识 多个对象以逗号隔开
    */
-  // public List<MembershipAuthorizationScopeObject> GetAuthorizationScopeObjects(String entityId, String authorityName) {
-  //  return provider.GetAuthorizationScopeObjects(entityId, authorityName);
-  // }
-  ///#endregion
+  @Override
+  public void bindAuthorizationScopeByAuthorizationObjectIds(String authorizationObjectType,
+    String authorizationObjectId, String authorityName, String entityIds) {
+    MembershipManagement.getInstance().getAuthorizationObjectService().bindAuthorizationScopeByAuthorizationObjectIds(
+      DATA_SCOPE_TABLE_NAME,
+      authorizationObjectType,
+      authorizationObjectId,
+      authorityName,
+      entityIds,
+      KernelContext.parseObjectType(ApplicationMenu.class));
+  }
+
+  /**
+   * 查询实体对象的授权信息
+   *
+   * @param entityId 实体标识
+   * @param authorityName 权限名称
+   */
+  @Override
+  public List<AuthorizationScope> getAuthorizationScopes(String entityId, String authorityName) {
+    return MembershipManagement.getInstance().getAuthorizationObjectService().getAuthorizationScopes(
+      DATA_SCOPE_TABLE_NAME,
+      entityId,
+      KernelContext.parseObjectType(ApplicationMenu.class),
+      authorityName);
+  }
 
   // -------------------------------------------------------
   // 权限
@@ -398,40 +380,21 @@ public class ApplicationMenuServiceImpl implements ApplicationMenuService {
    * 绑定 SQL 查询条件
    */
   private String bindAuthorizationScopeSql() {
-    if (AppsSecurity.isAdministrator(KernelContext.getCurrent().getUser(),
-      AppsConfiguration.APPLICATION_NAME)) {
+    if (AppsSecurity.isAdministrator(KernelContext.getCurrent().getUser(), APPLICATION_NAME)) {
       return "";
     } else {
-
       String accountId = KernelContext.getCurrent().getUser() == null ?
         UUIDUtil.emptyString() : KernelContext.getCurrent().getUser().getId();
 
       return String.format(" ("
-        + "(T.id IN ( "
+        + "(t.id IN ( "
         + "   SELECT entity_id "
-        + "     FROM view_authobject_account View1, application_menu_scope Scope"
-        + "    WHERE View1.account_id = '%s'"
-        + "      AND View1.authorization_object_id = Scope.authorization_object_id"
-        + "      AND View1.authorization_object_type = Scope.authorization_object_type "
+        + "     FROM view_authobject_account view1, application_menu_scope scope"
+        + "    WHERE view1.account_id = '%s'"
+        + "      AND view1.authorization_object_id = scope.authorization_object_id"
+        + "      AND view1.authorization_object_type = scope.authorization_object_type "
         + " GROUP BY entity_id)) "
         + ") ", accountId);
     }
-  }
-
-  /**
-   * @param id 角色
-   */
-  @Override
-  public List<String> getMenusScopeByRoleId(String id) {
-    List<String> result = new ArrayList<>();
-    List findResult = provider.getMenusScopeByRoleId(id);
-    if (findResult != null && findResult.size() > 0) {
-      for (Object find : findResult) {
-        Map<String, String> map = (Map<String, String>) find;
-        String entity_id = map.get("entity_id").toString();
-        result.add(entity_id);
-      }
-    }
-    return result;
   }
 }
